requirements.txt
bidict==0.22.0
Flask==2.2.2
python-engineio==4.3.4
python-socketio==5.7.2
waitress==2.1.2
youtube-dl==2021.12.17
